+++ internal/app/forklift/workspace.go
@@ -39,16 +39,16 @@ func LoadWorkspace(dirPath string) (*FSWorkspace, error) {
 }
 
 func (w *FSWorkspace) GetCurrentPalletPath() string {
-	return path.Join(w.getDataPath(), currentPalletDirName)
+	return path.Join(w.GetDataPath(), currentPalletDirName)
 }
 
-func (w *FSWorkspace) getDataPath() string {
+func (w *FSWorkspace) GetDataPath() string {
 	return path.Join(w.FS.Path(), dataDirPath)
 }
 
 func (w *FSWorkspace) GetCurrentPallet() (*FSPallet, error) {
-	if err := EnsureExists(w.getDataPath()); err != nil {
-		return nil, errors.Wrapf(err, "couldn't ensure the existence of %s", w.getDataPath())
+	if err := EnsureExists(w.GetDataPath()); err != nil {
+		return nil, errors.Wrapf(err, "couldn't ensure the existence of %s", w.GetDataPath())
 	}
 	return LoadFSPallet(w.FS, path.Join(dataDirPath, currentPalletDirName))
 }